+++ packages/halo/src/react/__specs__/index.spec.tsx
@@ -1,9 +1,9 @@
 import { render } from '@testing-library/react'
 import React from 'react'
 
-import * as vars from '../../vars'
+import * as vars from '../../vars/index.js'
 
-import Halo from '..'
+import Halo from '../index.js'
 
 describe('Halo', () => {
   describe('.gapSizes', () => {
